books/2015/pawga: init cljs (#47)
Since [Heart of Clojure], I am fully convinced that [shadow-cljs] is the
right way to manage ClojureScript in 2024. Still, all of my own apps are
stuck on [figwheel-main] at this point, because of inertia and lack of
time.

[Heart of Clojure]: https://2024.heartofclojure.eu
[shadow-cljs]: https://github.com/thheller/shadow-cljs
[figwheel-main]: https://figwheel.org

I am taking this project as an opportunity to dip my toes in the
shadow-cljs world. For this PR, what I've done is:

1. Skim the beginning of the shadow-cljs Quick Start.
2. Added NodeJS to `shell.nix`. (The `shadow-cljs` CLI tool is a node
   program.)
3. Added a JDK to `shell.nix`. (shdaow-cljs still calls out to the
   ClojureScript compiler, and, for advanced optimizations, to the
   Google Closure compiler; in both cases it needs a JVM.)
4. Ran the magic command `npx create-cljs-project cljs`, the results of
   which are the `cljs` folder added in this PR. (`npx` is provided by
   the NodeJS distribution.)

There's no ClojureScript code yet. We'll get there.