books/2015/pawga/ch3/cljs: handle victory conditions
This concludes my code exploration of Chapter 3. I'll move on to Chapter
4 now.

I suppose it's a matter of taste, but I find the wiring of components
togeher quite a bit easier with the ClojureScript approach, where I can
rely on message passing using immutable messages and concurrent,
synchronous code. The top-level of the project is a bit of a mess imo in
both JS versions, when trying to wrire together the core game loop with
click and resize events.