books/2015/pawga/ch3/cljs: add event loop (#84)
Now, we start doing things that are harder to do in pure JavaScript.

Note that we're not trying to get things actually happening in parallel
here, or to gain performance. This is about a better way to organize
code: instead of the pretty messy way we wire things together in the JS
version, here we can have nicely decoupled processes sending messages to
each other.