books/2015/pawga/ch4: add model events
Completely disconnected from everything at this point. I was worried the
diff may be too big otherwise. The idea is to simulate a separate model
loop that:

- Maintains the state of the game.
- Collects user input and updates the model accordingly.

I'm thinking for now this "game" will just be a knight moving on arrow
keys. So on right-arrow down, the model should emit a "animodel" that is
a function of time to the knight having moved, and on right-arrow up,
the model should emit a new animodel where the knight does not move
anymore, but has the same final position as displayed.

So this stub here is simulating a user that waits for a second when the
page loads, then presses the right arrow key for one second and then
releases it.