books/2015/pawga/ch5/cljs: almost working movement (#142)
This pretty much works as expected (except for the left-facing
direction), except for an unexpected issue: the knight starts vibrating
after about half a second.

What's happening is that my mental model so far was keydown means the
key was just pressed, and keyup means it was released.

It looks like the browser model is:

1. On key press, send a keydown event.
2. If the key is still pressed after about half a second, start sending
   a stream of keydown events in very short succession.

Which, I guess, is the age-old behaviour of keeping a key pressed to
repeat the same letter many times. It's not really what I want here
though.