aoc/2024/13: debug showed something
I get output that looks like:

```clojure
[44 16 51 75 2188 2828 [8 36] [9 36] 2232 2844]
[12 22 96 38 10188 5706 [97 94] [98 94] 10200 5728]
[13 62 79 85 6430 9367 [51 73] [52 73] 6443 9429]
[83 19 94 92 7904 5122 [42 47] [43 47] 7987 5141]
[24 78 87 80 1914 5004 [58 6] [59 6] 1938 5082]
[61 11 71 61 6710 4150 [39 61] [40 61] 6771 4161]
[60 13 97 94 5471 2864 [54 23] [55 23] 5531 2877]
[39 17 43 89 3593 2339 [80 11] [81 11] 3632 2356]
[16 22 82 27 6482 2653 [31 73] [32 73] 6498 2675]
[17 65 99 97 5914 9662 [80 46] [81 46] 5931 9727]
[14 37 97 44 4964 2926 [22 48] [23 48] 4978 2963]
[26 56 93 60 647 692 [7 5] [8 5] 673 748]
[15 34 97 59 4151 3296 [31 38] [32 38] 4166 3330]
[73 11 78 68 12004 7096 [64 94] [65 94] 12077 7107]
[12 42 95 67 6954 6816 [57 66] [58 66] 6966 6858]
```

so it looks like I can get into a situation where a is one too high, and
then I get to the "both are too big" condition and end there, rather
than try to reduce a one more time. I'll fix this ad-hoc and see if it
works.