aoc/2024/20: remove pq as we do all of it anyway
Also, the comparison function we give to the PriorityQueue does not
work: we compare `:cost` applied to vectors, so we always end up
comparing `nil` to `nil`.

This has virtually no impact on performance (~60s -> ~60s), but the
resulting code is slighlty simpler. Or at least not misleading: it's not
obvious at first glance that the comparison is meaningless, and its
existence could mislead on the format of the entries we're considering.