gity/minj: skip toolchain check as suggested
Following the suggestion from the previous output, I added a flag to
skip the toolchain check. This produces a warning:

```
Warning: The option '-H:-CheckToolchain' is experimental and must be enabled via '-H:+UnlockExperimentalVMOptions' in the future.
```

but otherwise seems to succeed, insofar as it does build a native
executable. However, running that executable produces the following
error:

```
$ ./minimalswingapp
Exception in thread "main" java.lang.UnsatisfiedLinkError: Can't load library: awt | java.library.path = [.]
        at org.graalvm.nativeimage.builder/com.oracle.svm.core.jdk.NativeLibraries.loadLibraryRelative(NativeLibraries.java:141)
        at java.base@24.0.1/java.lang.ClassLoader.loadLibrary(ClassLoader.java:100)
        at java.base@24.0.1/java.lang.Runtime.loadLibrary0(Runtime.java:822)
        at java.base@24.0.1/java.lang.System.loadLibrary(System.java:1663)
        at java.desktop@24.0.1/java.awt.Toolkit.loadLibraries(Toolkit.java:1293)
        at java.desktop@24.0.1/java.awt.Toolkit.initStatic(Toolkit.java:1318)
        at java.desktop@24.0.1/java.awt.Toolkit.<clinit>(Toolkit.java:1299)
        at java.desktop@24.0.1/java.awt.EventQueue.invokeLater(EventQueue.java:1257)
        at java.desktop@24.0.1/javax.swing.SwingUtilities.invokeLater(SwingUtilities.java:1415)
        at MinimalSwingApp.main(MinimalSwingApp.java:9)
        at java.base@24.0.1/java.lang.invoke.LambdaForm$DMH/sa346b79c.invokeStaticInit(LambdaForm$DMH)
```

This is the same issue I had before, in the parent folder, and the one
I've spent many hours scrounging Google for. I was hoping Gemini would
have some insight into this.

Its suggestion is that perhaps not using the Apple compiler is the root
of the issue: skipping the toolchain check allows the build to proceed,
but fails to build things "the right way", because Swing/AWT is so
deeply integrated with the host OS that it really does need the specific
Apple compiler and tools and the ones provided by Nix just aren't up to
snuff.

I don't know how much I buy that, but I had no alternative theory and it
sounded plausible, so that's the direction I went it.