fs/dups: try to make it parallel
It takes a while to process an item, but it also takes me a while to
think about what to do with an item, so hopefully this way me and the
computer can think at the same time rather than taking turns.

It looks like this lost a lot of the laziness, and also it crashes.