fs/dups: actually make it faster this time
So it turns out after adding debug logs it became clear why it was so
slow: I have a "sparsebundle" on one of my drives, which is implemented
as a bunch of very large files of exactly the same size. So with the
previous code I was reading and hashing all of those files in their
entirety.

Seeing that, I thought to myself: hashing is cool and all, but I really
only want to compare two files, so maybe I don't need to hash entire
files, and I can stop at the first bit that differs instead?

That sounds nice when you're thinking of a pair of files, but reading an
arbitrary number of files in parallel is unwieldy. So instead I went for
this idea of "let's see if the first chunk is the same" as a first
filter.

Runtime to first result: 1h15m -> 16m20s.