p/brc: translate official generator
The [original setup][0] was meant for a specific event at a point in
time and does not really make it easy to integrate (i.e. there is no
library I can rely on), and I don't really want to go through the
trouble of forking on GitHub.

This is a translation of [this specific file][1] as the instructions
seem to say this is what is meant to be used. No clue what the other
`CreateMeasurementsX.java` files are for.

[0]: https://github.com/gunnarmorling/1brc/tree/main
[1]: https://github.com/gunnarmorling/1brc/blob/db064194be375edc02d6dbcd21268ad40f7e2869/src/main/java/dev/morling/onebrc/CreateMeasurements.java