proglang: implement function application (#399)
This seems to sort of work, but environment management is not clear
enough in my mind that I'd be confident this is fully working.

And by "working" here I mean for the cases it covers; as of right now
it's wholly unclear to me how to manage the commented case, which
currently does not work. Specifically: how to handle mutations in the
captured environment. I mean, I can see options if I'm willing to use
mutation myself in my environments, but... I'd prefer not to do that, I
think?