proglang: thread global memory through (#405)
This is obviously not doing anything yet as `mem` is never read or set.
But it's getting threaded through everywhere, which is a nice first
step.

The goal is to have a system where:

- Environments are "per function", or more precisely per function call,
  i.e. closures remember the environment they were created in, and
  function application creates a new environment.
- Environments are maps from identifier to "memory location".
- The global memory map maps "memory location" to value.

I'm still not entirely sure what "memory location" should be, though the
biggest issue I'm seeing right now is garbage collection in the global
memory. With the current approach (env is a map of identifier to atom),
values are released by the host GC as soon as the corresponding
environment gets discarded. With this new approach, I lose that: the
environment would still be discarded, but the entry in the global memory
map wouldn't. [WeakHashMaps] may help, I suppose, though I think I'd
rather try out my own GC.

[WeakHashMaps]: https://docs.oracle.com/javase/8/docs/api/java/util/WeakHashMap.html