gity: remove 'clojure' variable (#310)
I still can't get it to work. I spent the next ~12 hours trying to get a
simple "hello world" Swing app in pure Java to work, with the latter
third of that making extensive use of Gemini's help, and really wasn't
able to figure out any way to make it work.

I eventually found [this ticket][0], which strongly suggests AWT (the
underpinnings of Swing) is simply not supported by GraalVM on macOS.

I also could not make it work with the [Liberica][1] GraalVM
distribution, which is supposed to have specific additional support for
Swing on macOS, and still couldn't get it to work.

I'll make a couple more commits to document what I've tried, but my
current thinking is to give up on Swing. People seem to report
marginally better experiences with JavaFX, so maybe that's an option.
I'm also starting to consider Electron-like approaches. Or just giving
up on native entirely, but that would make me a little bit sad.

[0]: https://github.com/oracle/graal/issues/4921
[1]: https://bell-sw.com/liberica-native-image-kit/