proglang: replace mdo with monad (mostly)
It turns out I lost a feature in the `monad` form by trying to be too
strict in my verification: `monad` only accepts a single symbol, whereas
`mdo` does not care and does not check, so anything that can go in the
arg vector of `fn` works.

And it turns out I've used that, so I have two forms that cannot be
easily translated to `monad` at this point.