proglang: BROKEN seemingly working parallel print
In the print output, `5` is printed by `t1`, then `2` and `3` are
printed by `t2`, then `11` is printed by `t1`. So, concurrent execution.

Yay.

I can't shake the feeling that my step function is "losing" `i-f`,
though. I can't come up with a good explanation as to why that's not
causing any issue.