proglang: BROKEN more generic way to handle effects
In this context, effects are all the monadic operations that are neither
`:pure` nor `:bind`, so in this case specifically `:print`. But writing
it this way means I can resolve all of the effects once, at the
"top-level" of `step`, rather than redoing them in the `bind` case as I
did previously.

The caveat of course is that effects need to eventually reach `:pure`
(or, I suppose, `:bind`, though that seems less likely) after successive
applications of `step`. I believe that's a fair assumption.