g/g-1: unblur ghost svg on Safari
For whatever reason, on (my) Safari the ghost gets blurry as soon as the
`translateZ` operation is applied. This is copy-pasting the `cgv-entity`
class's `transform` operation from [css-gameview], renaming it
`-webkit-transform` so as to not affect other browsers, and removing the
`translateZ` component.

The main downside of this approach is that I am no longer able to
translate along the Z axis, which I don't think I'd wanted to do in a 2d
game anyway.

[css-gameview]: https://github.com/chr15m/css-gameview