fs/dups: faster comparison by doing less work (#368)
According to a naive measurement, the time for running

```
lein run dups show 10
```

went down from 15 minutes to 9 minutes with this change. Anecdotally,
the program now only needs to read 2.4GB instead of 22.5GB to arrive at
that result. It looks like there are quite a few pairs of big files that
happen to have the same first 64k bytes, I guess.