proglang: use vatch (#440)
Well, this was a lot harder to debug than I expected. There is a subtle
difference in semantics between clonad's `match` and `vatch`: `vatch`
explicitly checks that the number of elements matches, whereas `match`
does not care.

Debugging this was made harder by the fact that there is no error when
you fall through a `vatch`; it just returns `nil`.