proglang/lisp: ast tests
This may not look like much, but let's take a step back and look at what
this means: we have a feature-full AST representation, that allows us to
express all of the features we have on the main proglang. This is
achieved with a 10-lines parser, as opposed to the 64 lines for the
equivalent in main.