proglang: BROKEN separate, simple parallel monad (#455)
This is a direct translation of the Java code found [here], though the
idioms are different enough that it's not necessarily an _obvious_
translation. I think (/hope) that doing this in a simpler example like
that will help me bring back the insight (do the inner bind) back to the
more complicated one.

Of course, right now there's nothing showing that it's parallel as we
only get the final results; I'll add that next, by adding a print. That
will also give me some global state to manage across threads.