proglang: BROKEN remove special handling of `:bind` (#459)
This makes sense. If I want to run `[:bind mv f]` step-wise, either `mv`
has been reduced to a `:pure` value and I can apply `f` to it, ir it's
not and I need to `step` the internal `mv` first.

Not sure why that felt so murky yesterday. It seems so clear to me now.