proglang: works again (#460)
I knew there was a nice, elegant solution in there somewhere. Not sure
why I did not see it yesterday. It seems so simple and obvious now.

So, we're back to a working (insofar as my tests pass) single-threaded
interpreter. What did we gain?

Well, for starters, it's now about 10x slower on my one benchmark, which
makes the gc overhead negligible. So that's a win, I guess?

More importantly, we are now explicitly progressing the main thread one
step at a time, which means we are now in a position to interleave that
with other threads.