proglang/lisp: introduce a reader
The traditional Lisp architecture is to have a separate reader. We're
cheating a bit here (well, not yet, but planning to) in that instaparse
already gives us data structures, but we're going to move towards having
`m-eval` work over values, rather than parse trees. In practice, for
now, that mostly means it'll switch to keywords with `v` as a prefix.

Here, though, the main reason I'm doing this is because I want to
implement `let` with a Clojure-like structure, meaning I need vectors,
so I wanted to test vectors on their own. Which lead me to realize the
distinction I'd started to make between `:int` and `:v/int` does not
really work long-term. So turning `:int` into `:v/int` is now going to
be the job of the reader (`read-form`), rather than `ml-eval`.