proglang/lisp: rewrite eval in terms of reader
It's not a huge gain, but adding this intermediate level of abstraction
highlights the fact that `eval` really only cares about symbols and
lists, and leaves the rest unchanged.

Well, actually, that's not entirely correct. Writing this, I just
realized as of this commit we would fail to evaluate `[(+ 1 2) 3]`.
Anyhoo.

The biggest disappointment of this piece of work is my realization that
I cannot namespace the keywords representing monadic values, because all
of the `:bind` (and some of the `:pure`) values are generated by
`monad`.