proglang/lisp: introduce a monad (#499)
Hopefully this time the gain is a bit more visible: we do get more
pieces (`m-run` and `m-eval` vs. just `m-eval`), but each piece is, if
we "believe" in monads and are able to read `monad` calls, simpler: the
threading through of `state` through `reduce` has disappeared from
`m-eval`. Intead of "doing" the interpretation, `m-eval` now returns a
plan for later interpretation by `m-run`, in a sense.

`m-run` at this point encodes two "operations": sequencing, threading
state, in `:bind`, and failure, through `:assert`.