proglang/lisp: self-recursive functions
This is perhaps a complicated way to get this working, but the more
straightforward way (as used in the Python version) requires a mutable
environment, so I guess it's time for me to introduce a memory again.

Also, that `replace-sym` function is going to be a pain moving forward,
if I keep it, I think.