aoc/2025/10: change of plans
Approaching this as a search is not working out. I'm going to completely
change my approach here and look at the problem as a system of equations
instead. That should be a small, finite number of steps per machine, and
thus be blazing fast. I feel like I coded that for a previous year and
it was a bit annoying, though. We'll see.